Refactor video context to separate provider
